CHGSubcontext
This script was used to parse and plot - CHG subcontext analysis data submitted for publication in Martinho et al 2021. Bisulfite sequencing data was extracted by a R script provided with the maniscrip generating multiple files. This script combines the output into a single table and plots the results for 7 genomic regions and two biological samples "WT" (cv M82) and "sulf".